Fixed telemetry
