+++ radio/src/telemetry/flysky_nv14.h
@@ -21,43 +21,7 @@
 #ifndef __FLYSKY_NV14_H__
 #define __FLYSKY_NV14_H__
 
-enum FlySkySensorType_E {
-  FLYSKY_SENSOR_RX_VOLTAGE,
-  FLYSKY_SENSOR_RX_SIGNAL,
-  FLYSKY_SENSOR_RX_RSSI,
-  FLYSKY_SENSOR_RX_NOISE,
-  FLYSKY_SENSOR_RX_SNR,
-  FLYSKY_SENSOR_TEMP,
-  FLYSKY_SENSOR_EXT_VOLTAGE,
-  FLYSKY_SENSOR_MOTO_RPM,
-  FLYSKY_SENSOR_PRESSURE,
-  FLYSKY_SENSOR_GPS
-};
-typedef struct FLYSKY_GPS_INFO_S {
-  uint8_t position_fix;
-  uint8_t satell_cnt;
-  uint8_t latitude[4];
-  uint8_t longtitude[4];
-  uint8_t altitude[4];
-  uint8_t g_speed[2];
-  uint8_t direction[2];
-} gps_info_t;
-typedef struct FLYSKY_SENSOR_DATA_S {
-  uint8_t sensor_type;
-  uint8_t sensor_id;
-  uint8_t voltage[2];
-  uint8_t signal;
-  uint8_t rssi[2];
-  uint8_t noise[2];
-  uint8_t snr[2];
-  uint8_t temp[2];
-  uint8_t ext_voltage[2];
-  uint8_t moto_rpm[2];
-  uint8_t pressure_value[2];
-  gps_info_t gps_info;
-} rx_sensor_t;
-
 void flySkyNv14SetDefault(int index, uint8_t id, uint8_t subId, uint8_t instance);
-void flySkyNv14ProcessTelemetryPacket(const uint8_t * ptr, uint8_t SensorType );
+void flySkyNv14ProcessTelemetryPacket(uint8_t * ptr, uint8_t len );
 
 #endif
